local: add local build help and github automation (!UPSTREAM)
Local only changes for our githib repo.
Changes to:
* make building easier for developers
* respond to local pull-requests
* automate build & publish for pushes and local PRs

Individual changes done by Alex and squashed by Bill.

github: tweak PR template for virtio-msg (!UPSTREAM)

While we are working on the virtio-msg specification in the repository
lets not confuse people by using the canned response for the upstream
repository.

REVISION: set to 1.5 working draft as a base (!UPSTREAM)

* Makefile: add some simple make automations

This is just a helpful shortcut, especially when editing the documents
within an IDE which will offer up make targets to build the documents.

make-setup-generated: optionally add GIT metadata

This isn't a full conversion to git based metadata but it allows for
local builds to add git commit and tree status to the final built
product. The normal formal build process is unaffected and uses
manually updated VERSION, REVISION and REVISION-DATE metadata.

github: add basic test build

github: add deploy step

This pushes the PDF as a release artefact.

github: add an update step before installing texlive

Otherwise you run into issues of the upstream has updated packages
since the last snapshot.

github: use a different release/publish pipeline

Having a draft means that the final asset won't be visible outside the
project which is what we want. This entails tagging a release each
time we run the steps.

github: only run the test phase on PRs

The release is done on pushes.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

github: use dedicated latex image

Installing all the latext tools from scratch takes some time and can
time out. Lets use an image that already has all the bits. The
texlive-full image is based on Alpin and we need to install a few
helper tools for build.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

github: add liberation font

We don't need it for the HTML but the full PDF needs fonts to render
properly.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

github: give tag and pdf names unique to owner fork and branch

The job number is not unique for different forks and this causes issues
with the tags which are global. Use the repo owners github name in the
tag and the branch as well.  Still include the job number to make unique.

Also change the pdf file name so we know where it came from.

Signed-off-by: Bill Mills <bill.mills@linaro.org>