Add a more information in Error message
Define a Message ID field in error message to carry the ID of the
message that triggered an error.
Define a Data Type field in Error message and a value STRING (1) so that
the rest of the payload of the message can be used as NULL terminating
string that can be printed.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>