virtio-msg: Add virtio-msg, a message based virtio transport layer
Add a new transport layer that is based on messages.

This transport layer still uses virtqueues as the other transport layers do
but implements transport layer operations by sending and receiving messages
instead of the "MMR" reads and writes used in virtio-mmio and virtio-pci.

This transport is useful when the device and driver are both implemented in
software but the trap and emulate operations of virtio-mmio and virtio-pci
can not be used.

This transport is intended to be used in many situations, including:
* between a host processor and its co-processors
* between two different systems (not SMP) connected via PCIe
* between normal and secure worlds
* host to vm
* vm to vm

This is an RFC and not yet intended to be merged. There are multiple
know issues including not conforming to virtio spec standards.

Signed-off-by: Bill Mills <bill.mills@linaro.org>
Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@amd.com>
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>