Update the message size definition
Update the message size definition to be clearer:
- minimum message size is 8 bytes (header and empty payload)
- minimum message size that a transport must support is 44 bytes
  (virtqueue set message)

Add a sentence to explain how size negocation works.

Modify get/set feature to use 32 bit blocks instead of allowing to
request any bit offest and size.

Modify bus message to discover devices to also include a number of
device IDs requested in the message (instead of relying on the message
size) to be more coherent with the other messages.

Add a mention in what a bus can do for restricting size and protocol
version that are negociated to restrict them to what it can handle.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>