virtio-msg: Eliminate the terms frontend and backend.
The virtio spec does not use the terms frontend and backend and
neither should we.  Replace frontend with driver and backend with device.

Frontend is easy to replace with driver and should be understood by all.

The term backend is not defined or used the same in industry.
Normally it means the whole device side implementation including
the virtual device, the bridge to a real device, and the driver for the
real device.  Here we are only interested in the virtual device
implementation.

At the spec level the terms "driver" and "device" are collective term for
all the software on that side and would include the device type code,
the common virtio-msg transport level and the virtio-msg bus specific level.

If we need to specifically call out an action from the bus level, use the
words "driver side bus" and "device side bus" respectively.

Signed-off-by: Bill Mills <bill.mills@linaro.org>