virtio-msg: Make it clear that no response is possible for EVENT_ etc.
Replace wording like "does not require a response" with.
"there is no response".

The first wording makes it sound like a response is possible but optional.
The second wording makes it sound like no response is possible which is
what we want.

The sender needs to know ahead of time if a response is expected.  So making
it optional has not purpose.  Eliminate variation and make it clear.

Signed-off-by: Bill Mills <bill.mills@linaro.org>