virtio-msg: notifications are mandatory from common transport POV
Change the language to show that the notification messages always exist
between the common transport layer and the bus implementation.

Show that the bus implementation has the choice of sending the event
messages in band, converting them to OoB events, or (last resort)
dropping them and reinventing them on the other side via pooling.

Signed-off-by: Bill Mills <bill.mills@linaro.org>