virtio-msg: handle device reset according to existing spec
Device initiated reset is done by setting the DEVICE_NEEDS_RESET bit in
the device status.  The driver then takes action.

Also make it clean that the driver only need to re-init the device if it
wants to use it again.  It is valid to leave it in the reset state if not.

Signed-off-by: Bill Mills <bill.mills@linaro.org>