virtio-msg: EVENT_AVAIL make "next wrap" a single bit like other transports
In other transports the "next wrap" field is a single bit and mapped to the
MSB of a 16 bit value.  We don't need a whole 32 bit value just for the
wrap value and having it this way causes confusion with how to set it.

Although we have a 32 bit value do the same as other transports and map the
wrap bit to the MSB and use the rest as the index.

Signed-off-by: Bill Mills <bill.mills@linaro.org>