virtio-msg: Fix virtqueue index value for "all virtqueues"
The hex value for this case was always intended to be "all 1s" but was
mis-specified before.  This fixes that.

This also makes it clear that for the EVENT_AVAIL use case, this would
typically be used on the device side.  (The driver side always knows which
virtqueue was updated as it originates this event.)

Signed-off-by: Bill Mills <bill.mills@linaro.org>