github: give tag and pdf names unique to owner fork and branch
The job number is not unique for different forks and this causes issues
with the tags which are global. Use the repo owners github name in the
tag and the branch as well.  Still include the job number to make unique.

Also change the pdf file name so we know where it came from.

Signed-off-by: Bill Mills <bill.mills@linaro.org>