virtio-msg: 8 byte header with unique identifier
Move the header to be 8 bytes and include a unique identifier that can
be use to correlate request to responses.

Rename the msg_id field to msg_op as having a msg_id and msg_uid field
and using identifiers in 2 contexts would have been unclear.

Rework ID to OP and change wording message identifer to message
operation to be coherent in the rest of the specification.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>