virtio-msg: Relax error handling in the bus
Relax error handling in the virtio-msg bus and introduce wording to
allow the virtio-msg bus to report error to the virtio-msg transport if
a request could not be transmited or a malformed response was received.

This prevents the need to generate dummy responses in the bus and make
the specification coherent with the current Linux PoC implementation.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>