virtio-msg: remove redundant length fields
Remove length fields that provide no extra information (as the message
header has the full message length).

This changes:
* SET_FEATURES
* SET_CONFIG

SET_CONFIG rejection is harder to explain now so expand the overview
section and describe both the accepted and rejected cases.

NOTE: The number of blocks/bytes is still required in the GET_ messages
      so this change makes the GET_ and SET_ message layouts different.

Signed-off-by: Bill Mills <bill.mills@linaro.org>